Update the GTK+ 2.x GUI
Tthis stuff certainly ain't my best work ever, but I'm leaving the
structure of the code largely intact, as it still works perfectly.

I made a few cosmetic alterations, and updated it to add an error
dialog, and to work around some deprecated functions; but otherwise it's
largely the same as it was when I wrote it years ago.

It's interesting how the way you code changes over time.